+++ modules/spring-web/src/test/kotlin/com/esprito/spring/web/completion/kotlin/UastMockMvcUrlReferenceContributorTest.kt
@@ -249,4 +249,59 @@ class UastMockMvcUrlReferenceContributorTest : EspritoKotlinLightTestCase() {
         val multiResolve = ref!!.multiResolve(true)
         assertEquals(0, multiResolve.size)
     }
+
+    fun testTemplate() {
+        myFixture.copyFileToProject("ProductController.kt")
+        myFixture.configureByText(
+            "ProductControllerTest.kt", """
+            import org.springframework.test.web.servlet.request.MockMvcRequestBuilders
+            
+            class ProductControllerTest {
+                fun justForTest() {
+                    MockMvcRequestBuilders.put("/product/someProdu<caret>ctId")
+                }
+            }
+        """.trimIndent()
+        )
+
+        val ref = file.findReferenceAt(myFixture.caretOffset) as? EspritoControllerMethodReference
+        assertNotNull(ref)
+        val multiResolve = ref!!.multiResolve(true)
+        assertEquals(1, multiResolve.size)
+        val resolvedElement = (multiResolve[0].element) as? PsiMember
+        assertNotNull(resolvedElement)
+        val memberName = resolvedElement?.name
+        val classFqn = resolvedElement?.containingClass?.qualifiedName
+        TestCase.assertEquals("ProductController#update", "$classFqn#$memberName")
+    }
+
+    fun testTemplateWithConst() {
+        myFixture.copyFileToProject("ProductController.kt")
+        myFixture.configureByText(
+            "ProductControllerTest.kt", """
+            import org.springframework.test.web.servlet.request.MockMvcRequestBuilders
+            
+            class ProductControllerTest {
+                fun justForTest() {
+                    MockMvcRequestBuilders.put("/prod<caret>uct/" + ID)
+                }
+
+                companion object {
+                    private const val ID = "someProductId"
+                }
+            }
+        """.trimIndent()
+        )
+
+        val ref = file.findReferenceAt(myFixture.caretOffset) as? EspritoControllerMethodReference
+        assertNotNull(ref)
+        val multiResolve = ref!!.multiResolve(true)
+        assertEquals(1, multiResolve.size)
+        val resolvedElement = (multiResolve[0].element) as? PsiMember
+        assertNotNull(resolvedElement)
+        val memberName = resolvedElement?.name
+        val classFqn = resolvedElement?.containingClass?.qualifiedName
+        TestCase.assertEquals("ProductController#update", "$classFqn#$memberName")
+    }
+
 }